add/remove album function

+++ src/main/java/com/zosia/zosia/http/album/response/album/AlbumService.java
@@ -10,9 +10,11 @@ import com.zosia.zosia.http.album.response.image.ImageRepository;
 import com.zosia.zosia.http.album.response.label.LabelRepository;
 import com.zosia.zosia.http.album.response.track.TrackRepository;
 import com.zosia.zosia.http.service.HttpService;
+import com.zosia.zosia.user.User;
+import com.zosia.zosia.user.UserRepository;
 import org.springframework.stereotype.Service;
 
-
+import java.util.Set;
 
 
 @Service
@@ -25,9 +27,11 @@ public class AlbumService {
 	private final TrackRepository trackRepository;
 	private final ArtistRepository artistRepository;
 	private final AlbumRepository albumRepository;
+	private final UserRepository userRepository;
 	
 	public AlbumService (HttpService httpService, MessageService messageService, ImageRepository imageRepository, LabelRepository labelRepository,
-						 TrackRepository trackRepository, ArtistRepository artistRepository, AlbumRepository albumRepository) {
+						 TrackRepository trackRepository, ArtistRepository artistRepository, AlbumRepository albumRepository,
+						 UserRepository userRepository) {
 		
 		this.httpService = httpService;
 		this.messageService = messageService;
@@ -37,6 +41,7 @@ public class AlbumService {
 		
 		this.artistRepository = artistRepository;
 		this.albumRepository = albumRepository;
+		this.userRepository = userRepository;
 	}
 	
 	public <T> T requestAlbumBuilder (Class<T> type, String id) throws JsonProcessingException {
@@ -46,14 +51,40 @@ public class AlbumService {
 		
 	}
 	
-	public void saveAlbum (String id) throws JsonProcessingException {
+	public void saveAlbum (String id, User user) throws JsonProcessingException {
 		
-		Album album = requestAlbumBuilder(Album.class, id);
 		
+		if (albumRepository.findById(Long.parseLong(id)).isEmpty()) {
+			Album album = requestAlbumBuilder(Album.class, id);
+			album.addUser(user);
+			albumRepository.save(album);
+		}
+		else {
+			Album a = albumRepository.findById(Long.parseLong(id)).get();
+			a.addUser(user);
+			albumRepository.save(a);
+		}
+	}
 	
-		albumRepository.delete(album);
+	public void deleteAlbum (String id, User user) throws JsonProcessingException {
 		
+		Album album = albumRepository.findById(Long.parseLong(id)).get();
+		album.removeUser(userRepository.findById(user.getId()).get());
+		albumRepository.save(album);
 		
+		if (album.getUsers().isEmpty()) {
+			album.removeArtist();
+			album.removeLabel();
+			albumRepository.save(album);
+			albumRepository.delete(album);
+			
+			artistRepository.findAll().forEach(artist -> {
+				if (artist.getAlbums().isEmpty()) {artistRepository.delete(artist);}
+			});
+			labelRepository.findAll().forEach(label -> {
+				if (label.getAlbums().isEmpty()) {labelRepository.delete(label);}
+			});
+		}
 	}
 	
 }
